runtime/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "factory-runtime"
version = "0.1.0"
description = "Local execution environment for a team of six persistent AI agents"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.42.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
]

[project.scripts]
factory = "factory_runtime.cli:main"
